core,graph,server: Implement the udpate_subgraph_price RPC method
Update the query price for a specified subgraph.
Use u64 for query price, this is assuming that the price will be in
gwei denominated integers on the contracts.

+++ graph/src/components/subgraph/registrar.rs
@@ -44,7 +44,7 @@ pub trait SubgraphRegistrar: Send + Sync + 'static {
     fn update_subgraph_price(
         &self,
         hash: SubgraphDeploymentId,
-        price: f32
+        price: u64
     ) -> Box<dyn Future<Item = (), Error = SubgraphRegistrarError> + Send + 'static>;
 
     fn update_all_prices(
